feat: implement utility function to compute cart items
